ns2/ns3: decode the second panel, and ns3y with it
Each Stage program stores two of everything, and the byte maps document one.
The Stage 3 body is 22 bytes of globals then two 263-byte panels; the Stage 2
is 23 bytes then two 249-byte slots. The second repeats the first, so panel A's
placements replicate at the stride: ns3f goes 308 -> 588 fields, ns2p 436 ->
838. Panel A's fields keep their names; the second block carries panel_b_ /
slot_b_.

The stride is not a guess. Three independent measurements agree: locating synth
preset content inside program bodies gives exactly two offsets 263 apart on the
Stage 3 and 249 on the Stage 2; the body arithmetic lands exactly
(548 = 2*263 + 22, 521 = 2*249 + 23); and shifting every typed selector by the
stride scores 0.00% out-of-table over 2640 reads where +260 and +267 score ~36%.
The corpus tripwire now covers the second block's selectors too.

That also settles ns3y, which the last pass left as a stub because the
documented offset did not reproduce. It is panel A's synth block at program
body 0x4f — five bytes before the docs' claim — and panel B's copy at 0x4f+263
is what made the two-panel structure visible in the first place. 58 fields,
sharing the program's own names and enums so a parameter reads the same either
side of the tag.

+++ crates/nord-format/tests/dispatch.rs
@@ -64,7 +64,7 @@ fn formats() -> Vec<(&'static str, u64, u32)> {
         (ns3::program::FORMAT, ns3::program::BODY_LEN as u64, 304),
         (ns3::live::FORMAT, ns3::program::BODY_LEN as u64, 304),
         (ns3::song::FORMAT, ns3::song::BODY_LEN, 300),
-        (ns3::synth::FORMAT, ns3::synth::BODY_LEN, 300),
+        (ns3::synth::FORMAT, ns3::synth::BODY_LEN as u64, 300),
         (ns3::settings::FORMAT, ns3::settings::BODY_LEN, 300),
         (ns4::program::FORMAT, ns4::program::BODY_LEN as u64, 313),
         (ns4::live::FORMAT, ns4::program::BODY_LEN as u64, 313),
